+++ braindecode/models/eegitnet.py
@@ -3,17 +3,9 @@
 # License: BSD-3
 import torch
 from torch import nn
+from einops.layers.torch import Rearrange
 
-from .modules import Ensure4d, Expression
-
-
-def _permute(x):
-    """Permute data.
-
-    Input dimensions: (batch, channels, time, 1)
-    Output dimiensions: (batch, 1, channels, time)
-    """
-    return x.permute([0, 3, 1, 2])
+from .modules import Ensure4d
 
 
 class _DepthwiseConv2d(torch.nn.Conv2d):
@@ -131,7 +123,9 @@ class EEGITNet(nn.Sequential):
 
         # ======== Handling EEG input ========================
         self.add_module(
-            "input_preprocess", nn.Sequential(Ensure4d(), Expression(_permute))
+            "input_preprocess", nn.Sequential(Ensure4d(),
+                                              Rearrange(
+                                                  "ba ch t 1 -> ba 1 ch t"))
         )
         # ======== Inception branches ========================
         block11 = self._get_inception_branch(